dvdplayer: Fix m_streaminfo updates from passthrough codec
Use the GetEncodedSampleRate() method instead of GetSampleRate() to
populate m_streaminfo, since the latter returns the passthrough frame
rate which is not relevant in the context of m_streaminfo.

Also, do not overwrite channel count and sample rate with zero after
opening a passthrough codec, instead keep the demuxer values. These are
only used internally for codec selection.

This fixes some unnecessary back-and-forth checking between regular
and passthrough codec when opening a stream due to wrongly detected
sample rate changes in certain situations.