Fix back button when filters are not loaded.
This was kind of annoying before. The first time you went into a
section it would not exit properly, you would have to exit twice. This
was because the sectionFilter was not loaded and the primaryFilter was
not properly initialised. Now instead we will assume that it’s ‘all’ if
it’s not loaded.