[RasPlex] Avoid Decoding / reencoding of textures when caching
PMS does the job for us and already provides proper Jpg/Png for us, so all we have to do is to sabe the stream to a file which is way more efficient than decoding the PMS stream and reencode it to a local file.

This brings a huge boost on RPI for image caching.

I have put that feature that we have been using for a couple weeks now in a Raspberry define for safety, but all plex client could benefit from this :)