GUI.OnScreensaverDeactivated does not follow JSON-RPC API definition
The announcement of `GUI.OnScreensaverDeactivated` does not follow the
defined API - it outputs `data` as a `boolean`, not `null`.  This
behaviour was introduced in ea9f2f01 to support a special case for CEC
devices, that must choose to act based on whether this notification was
triggered on the back of a shutdown.

As the rest of the API defines `data` to be of type `null` or `object`,
I've moved this boolean to a `shuttingDown` member on the data object,
rather than changing the API to be inconsistent for a special case.